separate children into children, children_ref, child, child_ref
