+++ bastion/src/children.rs
@@ -1,12 +1,14 @@
 //!
 //! Children are a group of child supervised under a supervisor
-//! Allows users to communicate with children through the mailboxes.
 use crate::broadcast::{Broadcast, Parent, Sender};
 use crate::callbacks::Callbacks;
+use crate::child::{Child, Init};
+use crate::child_ref::ChildRef;
+use crate::children_ref::ChildrenRef;
 use crate::context::{BastionContext, BastionId, ContextState};
-use crate::envelope::{Envelope, RefAddr};
-use crate::message::{Answer, BastionMessage, Message};
-use crate::path::{BastionPath, BastionPathElement};
+use crate::envelope::Envelope;
+use crate::message::BastionMessage;
+use crate::path::BastionPathElement;
 use bastion_executor::pool;
 use futures::pending;
 use futures::poll;
@@ -15,16 +17,10 @@ use futures::stream::{FuturesOrdered, FuturesUnordered};
 use fxhash::FxHashMap;
 use lightproc::prelude::*;
 use qutex::Qutex;
-use std::cmp::{Eq, PartialEq};
-use std::fmt::{self, Debug, Formatter};
+use std::fmt::Debug;
 use std::future::Future;
 use std::iter::FromIterator;
-use std::pin::Pin;
-use std::sync::Arc;
-use std::task::{Context, Poll};
-
-struct Init(Box<dyn Fn(BastionContext) -> Exec + Send + Sync>);
-struct Exec(Pin<Box<dyn Future<Output = Result<(), ()>> + Send>>);
+use std::task::Poll;
 
 #[derive(Debug)]
 /// A children group that will contain a defined number of
@@ -88,59 +84,6 @@ pub struct Children {
     // is received.
     pre_start_msgs: Vec<Envelope>,
     started: bool,
-}
-
-#[derive(Debug, Clone)]
-/// A "reference" to a children group, allowing to communicate
-/// with it.
-pub struct ChildrenRef {
-    id: BastionId,
-    sender: Sender,
-    path: Arc<BastionPath>,
-    children: Vec<ChildRef>,
-}
-
-#[derive(Debug)]
-pub(crate) struct Child {
-    bcast: Broadcast,
-    // The future that this child is executing.
-    exec: Exec,
-    // A lock behind which is the child's context state.
-    // This is used to store the messages that were received
-    // for the child's associated future to be able to
-    // retrieve them.
-    state: Qutex<ContextState>,
-    // Messages that were received before the child was
-    // started. Those will be "replayed" once a start message
-    // is received.
-    pre_start_msgs: Vec<Envelope>,
-    started: bool,
-}
-
-#[derive(Debug, Clone)]
-/// A "reference" to an element of a children group, allowing to
-/// communicate with it.
-pub struct ChildRef {
-    id: BastionId,
-    sender: Sender,
-    path: Arc<BastionPath>,
-}
-
-impl Init {
-    fn new<C, F>(init: C) -> Self
-    where
-        C: Fn(BastionContext) -> F + Send + Sync + 'static,
-        F: Future<Output = Result<(), ()>> + Send + 'static,
-    {
-        let init = Box::new(move |ctx: BastionContext| {
-            let fut = init(ctx);
-            let exec = Box::pin(fut);
-
-            Exec(exec)
-        });
-
-        Init(init)
-    }
 }
 
 impl Children {
@@ -636,720 +579,5 @@ impl Children {
         debug!("Children({}): Launching.", self.id());
         let stack = self.stack();
         pool::spawn(self.run(), stack)
-    }
-}
-
-impl ChildrenRef {
-    fn new(id: BastionId, sender: Sender, path: Arc<BastionPath>, children: Vec<ChildRef>) -> Self {
-        ChildrenRef {
-            id,
-            sender,
-            path,
-            children,
-        }
-    }
-
-    /// Returns the identifier of the children group this `ChildrenRef`
-    /// is referencing.
-    ///
-    /// Note that the children group's identifier is reset when it
-    /// is restarted.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// # use bastion::prelude::*;
-    /// #
-    /// # fn main() {
-    ///     # Bastion::init();
-    ///     #
-    /// let children_ref = Bastion::children(|children| {
-    ///     // ...
-    ///     # children
-    /// }).expect("Couldn't create the children group.");
-    ///
-    /// let children_id: &BastionId = children_ref.id();
-    ///     #
-    ///     # Bastion::start();
-    ///     # Bastion::stop();
-    ///     # Bastion::block_until_stopped();
-    /// # }
-    /// ```
-    pub fn id(&self) -> &BastionId {
-        &self.id
-    }
-
-    /// Returns a list of [`ChildRef`] referencing the elements
-    /// of the children group this `ChildrenRef` is referencing.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// # use bastion::prelude::*;
-    /// #
-    /// # fn main() {
-    ///     # Bastion::init();
-    ///     #
-    ///     # let children_ref = Bastion::children(|children| children).unwrap();
-    /// let elems: &[ChildRef] = children_ref.elems();
-    ///     #
-    ///     # Bastion::start();
-    ///     # Bastion::stop();
-    ///     # Bastion::block_until_stopped();
-    /// # }
-    /// ```
-    ///
-    /// [`ChildRef`]: children/struct.ChildRef.html
-    pub fn elems(&self) -> &[ChildRef] {
-        &self.children
-    }
-
-    /// Sends a message to the children group this `ChildrenRef`
-    /// is referencing which will then send it to all of its
-    /// elements.
-    ///
-    /// An alternative would be to use [`elems`] to get all the
-    /// elements of the group and then send the message to all
-    /// of them.
-    ///
-    /// This method returns `()` if it succeeded, or `Err(msg)`
-    /// otherwise.
-    ///
-    /// # Arguments
-    ///
-    /// * `msg` - The message to send.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// # use bastion::prelude::*;
-    /// #
-    /// # fn main() {
-    ///     # Bastion::init();
-    ///     #
-    ///     # let children_ref = Bastion::children(|children| children).unwrap();
-    /// let msg = "A message containing data.";
-    /// children_ref.broadcast(msg).expect("Couldn't send the message.");
-    ///
-    ///     # Bastion::children(|children| {
-    ///         # children.with_exec(|ctx: BastionContext| {
-    ///             # async move {
-    /// // And then in every of the children group's elements' futures...
-    /// msg! { ctx.recv().await?,
-    ///     ref msg: &'static str => {
-    ///         assert_eq!(msg, &"A message containing data.");
-    ///     };
-    ///     // We are only sending a `&'static str` in this
-    ///     // example, so we know that this won't happen...
-    ///     _: _ => ();
-    /// }
-    ///                 #
-    ///                 # Ok(())
-    ///             # }
-    ///         # })
-    ///     # }).unwrap();
-    ///     #
-    ///     # Bastion::start();
-    ///     # Bastion::stop();
-    ///     # Bastion::block_until_stopped();
-    /// # }
-    /// ```
-    ///
-    /// [`elems`]: #method.elems
-    pub fn broadcast<M: Message>(&self, msg: M) -> Result<(), M> {
-        debug!(
-            "ChildrenRef({}): Broadcasting message: {:?}",
-            self.id(),
-            msg
-        );
-        let msg = BastionMessage::broadcast(msg);
-        let env = Envelope::from_dead_letters(msg);
-        // FIXME: panics?
-        self.send(env).map_err(|err| err.into_msg().unwrap())
-    }
-
-    /// Sends a message to the children group this `ChildrenRef`
-    /// is referencing to tell it to stop all of its running
-    /// elements.
-    ///
-    /// This method returns `()` if it succeeded, or `Err(())`
-    /// otherwise.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// # use bastion::prelude::*;
-    /// #
-    /// # fn main() {
-    ///     # Bastion::init();
-    ///     #
-    ///     # let children_ref = Bastion::children(|children| children).unwrap();
-    /// children_ref.stop().expect("Couldn't send the message.");
-    ///     #
-    ///     # Bastion::start();
-    ///     # Bastion::stop();
-    ///     # Bastion::block_until_stopped();
-    /// # }
-    /// ```
-    pub fn stop(&self) -> Result<(), ()> {
-        debug!("ChildrenRef({}): Stopping.", self.id());
-        let msg = BastionMessage::stop();
-        let env = Envelope::from_dead_letters(msg);
-        self.send(env).map_err(|_| ())
-    }
-
-    /// Sends a message to the children group this `ChildrenRef`
-    /// is referencing to tell it to kill all of its running
-    /// elements.
-    ///
-    /// This method returns `()` if it succeeded, or `Err(())`
-    /// otherwise.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// # use bastion::prelude::*;
-    /// #
-    /// # fn main() {
-    ///     # Bastion::init();
-    ///     #
-    ///     # let children_ref = Bastion::children(|children| children).unwrap();
-    /// children_ref.kill().expect("Couldn't send the message.");
-    ///     #
-    ///     # Bastion::start();
-    ///     # Bastion::stop();
-    ///     # Bastion::block_until_stopped();
-    /// # }
-    /// ```
-    pub fn kill(&self) -> Result<(), ()> {
-        debug!("ChildrenRef({}): Killing.", self.id());
-        let msg = BastionMessage::kill();
-        let env = Envelope::from_dead_letters(msg);
-        self.send(env).map_err(|_| ())
-    }
-
-    pub(crate) fn send(&self, env: Envelope) -> Result<(), Envelope> {
-        trace!("ChildrenRef({}): Sending message: {:?}", self.id(), env);
-        self.sender
-            .unbounded_send(env)
-            .map_err(|err| err.into_inner())
-    }
-
-    pub(crate) fn path(&self) -> &Arc<BastionPath> {
-        &self.path
-    }
-
-    pub(crate) fn sender(&self) -> &Sender {
-        &self.sender
-    }
-}
-
-impl Child {
-    fn new(exec: Exec, bcast: Broadcast, state: Qutex<ContextState>) -> Self {
-        debug!("Child({}): Initializing.", bcast.id());
-        let pre_start_msgs = Vec::new();
-        let started = false;
-
-        Child {
-            bcast,
-            exec,
-            state,
-            pre_start_msgs,
-            started,
-        }
-    }
-
-    fn stack(&self) -> ProcStack {
-        trace!("Child({}): Creating ProcStack.", self.id());
-        let id = self.bcast.id().clone();
-        // FIXME: panics?
-        let parent = self.bcast.parent().clone().into_children().unwrap();
-        let path = self.bcast.path().clone();
-        let sender = self.bcast.sender().clone();
-
-        // FIXME: with_pid
-        ProcStack::default().with_after_panic(move || {
-            // FIXME: clones
-            let id = id.clone();
-            warn!("Child({}): Panicked.", id);
-
-            let msg = BastionMessage::faulted(id);
-            let env = Envelope::new(msg, path.clone(), sender.clone());
-            // TODO: handle errors
-            parent.send(env).ok();
-        })
-    }
-
-    fn id(&self) -> &BastionId {
-        self.bcast.id()
-    }
-
-    fn stopped(&mut self) {
-        debug!("Child({}): Stopped.", self.id());
-        self.bcast.stopped();
-    }
-
-    fn faulted(&mut self) {
-        debug!("Child({}): Faulted.", self.id());
-        self.bcast.faulted();
-    }
-
-    async fn handle(&mut self, env: Envelope) -> Result<(), ()> {
-        match env {
-            Envelope {
-                msg: BastionMessage::Start,
-                ..
-            } => unreachable!(),
-            Envelope {
-                msg: BastionMessage::Stop,
-                ..
-            } => {
-                self.stopped();
-
-                return Err(());
-            }
-            Envelope {
-                msg: BastionMessage::Kill,
-                ..
-            } => {
-                self.stopped();
-
-                return Err(());
-            }
-            // FIXME
-            Envelope {
-                msg: BastionMessage::Deploy(_),
-                ..
-            } => unimplemented!(),
-            // FIXME
-            Envelope {
-                msg: BastionMessage::Prune { .. },
-                ..
-            } => unimplemented!(),
-            // FIXME
-            Envelope {
-                msg: BastionMessage::SuperviseWith(_),
-                ..
-            } => unimplemented!(),
-            Envelope {
-                msg: BastionMessage::Message(msg),
-                sign,
-            } => {
-                debug!("Child({}): Received a message: {:?}", self.id(), msg);
-                let mut state = self.state.clone().lock_async().await.map_err(|_| ())?;
-                state.push_msg(msg, sign);
-            }
-            // FIXME
-            Envelope {
-                msg: BastionMessage::Stopped { .. },
-                ..
-            } => unimplemented!(),
-            // FIXME
-            Envelope {
-                msg: BastionMessage::Faulted { .. },
-                ..
-            } => unimplemented!(),
-        }
-
-        Ok(())
-    }
-
-    async fn run(mut self) {
-        debug!("Child({}): Launched.", self.id());
-        loop {
-            match poll!(&mut self.bcast.next()) {
-                // TODO: Err if started == true?
-                Poll::Ready(Some(Envelope {
-                    msg: BastionMessage::Start,
-                    ..
-                })) => {
-                    trace!(
-                        "Child({}): Received a new message (started=false): {:?}",
-                        self.id(),
-                        BastionMessage::Start
-                    );
-                    debug!("Child({}): Starting.", self.id());
-                    self.started = true;
-
-                    let msgs = self.pre_start_msgs.drain(..).collect::<Vec<_>>();
-                    self.pre_start_msgs.shrink_to_fit();
-
-                    debug!(
-                        "Child({}): Replaying messages received before starting.",
-                        self.id()
-                    );
-                    for msg in msgs {
-                        trace!("Child({}): Replaying message: {:?}", self.id(), msg);
-                        if self.handle(msg).await.is_err() {
-                            return;
-                        }
-                    }
-
-                    continue;
-                }
-                Poll::Ready(Some(msg)) if !self.started => {
-                    trace!(
-                        "Child({}): Received a new message (started=false): {:?}",
-                        self.id(),
-                        msg
-                    );
-                    self.pre_start_msgs.push(msg);
-
-                    continue;
-                }
-                Poll::Ready(Some(msg)) => {
-                    trace!(
-                        "Child({}): Received a new message (started=true): {:?}",
-                        self.id(),
-                        msg
-                    );
-                    if self.handle(msg).await.is_err() {
-                        return;
-                    }
-
-                    continue;
-                }
-                // NOTE: because `Broadcast` always holds both a `Sender` and
-                //      `Receiver` of the same channel, this would only be
-                //      possible if the channel was closed, which never happens.
-                Poll::Ready(None) => unreachable!(),
-                Poll::Pending => (),
-            }
-
-            if !self.started {
-                pending!();
-
-                continue;
-            }
-
-            match poll!(&mut self.exec) {
-                Poll::Ready(Ok(())) => {
-                    debug!(
-                        "Child({}): The future finished executing successfully.",
-                        self.id()
-                    );
-                    return self.stopped();
-                }
-                Poll::Ready(Err(())) => {
-                    warn!("Child({}): The future returned an error.", self.id());
-                    return self.faulted();
-                }
-                Poll::Pending => (),
-            }
-
-            pending!();
-        }
-    }
-
-    fn launch(self) -> RecoverableHandle<()> {
-        let stack = self.stack();
-        pool::spawn(self.run(), stack)
-    }
-}
-
-impl ChildRef {
-    fn new(id: BastionId, sender: Sender, path: Arc<BastionPath>) -> ChildRef {
-        ChildRef { id, sender, path }
-    }
-
-    /// Returns the identifier of the children group element this
-    /// `ChildRef` is referencing.
-    ///
-    /// Note that the children group element's identifier is reset
-    /// when it is restarted.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// # use bastion::prelude::*;
-    /// #
-    /// # fn main() {
-    ///     # Bastion::init();
-    ///     #
-    /// Bastion::children(|children| {
-    ///     children.with_exec(|ctx| {
-    ///         async move {
-    ///             let child_id: &BastionId = ctx.current().id();
-    ///             // ...
-    ///             # Ok(())
-    ///         }
-    ///     })
-    /// }).expect("Couldn't create the children group.");
-    ///     #
-    ///     # Bastion::start();
-    ///     # Bastion::stop();
-    ///     # Bastion::block_until_stopped();
-    /// # }
-    /// ```
-    pub fn id(&self) -> &BastionId {
-        &self.id
-    }
-
-    /// Sends a message to the child this `ChildRef` is referencing.
-    /// This message is intended to be used outside of Bastion context when
-    /// there is no way for receiver to identify message sender
-    ///
-    /// This method returns `()` if it succeeded, or `Err(msg)`
-    /// otherwise.
-    ///
-    /// # Argument
-    ///
-    /// * `msg` - The message to send.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// # use bastion::prelude::*;
-    /// #
-    /// # fn main() {
-    ///     # Bastion::init();
-    /// // The message that will be "told"...
-    /// const TELL_MSG: &'static str = "A message containing data (tell).";
-    ///
-    ///     # let children_ref =
-    /// // Create a new child...
-    /// Bastion::children(|children| {
-    ///     children.with_exec(|ctx: BastionContext| {
-    ///         async move {
-    ///             // ...which will receive the message "told"...
-    ///             msg! { ctx.recv().await?,
-    ///                 msg: &'static str => {
-    ///                     assert_eq!(msg, TELL_MSG);
-    ///                     // Handle the message...
-    ///                 };
-    ///                 // This won't happen because this example
-    ///                 // only "tells" a `&'static str`...
-    ///                 _: _ => ();
-    ///             }
-    ///
-    ///             Ok(())
-    ///         }
-    ///     })
-    /// }).expect("Couldn't create the children group.");
-    ///
-    ///     # let child_ref = &children_ref.elems()[0];
-    /// // Later, the message is "told" to the child...
-    /// child_ref.tell_anonymously(TELL_MSG).expect("Couldn't send the message.");
-    ///     #
-    ///     # Bastion::start();
-    ///     # Bastion::stop();
-    ///     # Bastion::block_until_stopped();
-    /// # }
-    /// ```
-    pub fn tell_anonymously<M: Message>(&self, msg: M) -> Result<(), M> {
-        debug!("ChildRef({}): Telling message: {:?}", self.id(), msg);
-        let msg = BastionMessage::tell(msg);
-        let env = Envelope::from_dead_letters(msg);
-        // FIXME: panics?
-        self.send(env).map_err(|env| env.into_msg().unwrap())
-    }
-
-    /// Sends a message to the child this `ChildRef` is referencing,
-    /// allowing it to answer.
-    /// This message is intended to be used outside of Bastion context when
-    /// there is no way for receiver to identify message sender
-    ///
-    /// This method returns [`Answer`](../message/struct.Answer.html) if it succeeded, or `Err(msg)`
-    /// otherwise.
-    ///
-    /// # Argument
-    ///
-    /// * `msg` - The message to send.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use bastion::prelude::*;
-    /// #
-    /// # fn main() {
-    ///     # Bastion::init();
-    /// // The message that will be "asked"...
-    /// const ASK_MSG: &'static str = "A message containing data (ask).";
-    /// // The message the will be "answered"...
-    /// const ANSWER_MSG: &'static str = "A message containing data (answer).";
-    ///
-    ///     # let children_ref =
-    /// // Create a new child...
-    /// Bastion::children(|children| {
-    ///     children.with_exec(|ctx: BastionContext| {
-    ///         async move {
-    ///             // ...which will receive the message asked...
-    ///             msg! { ctx.recv().await?,
-    ///                 msg: &'static str =!> {
-    ///                     assert_eq!(msg, ASK_MSG);
-    ///                     // Handle the message...
-    ///
-    ///                     // ...and eventually answer to it...
-    ///                     answer!(ctx, ANSWER_MSG);
-    ///                 };
-    ///                 // This won't happen because this example
-    ///                 // only "asks" a `&'static str`...
-    ///                 _: _ => ();
-    ///             }
-    ///
-    ///             Ok(())
-    ///         }
-    ///     })
-    /// }).expect("Couldn't create the children group.");
-    ///
-    ///     # Bastion::children(|children| {
-    ///         # children.with_exec(move |ctx: BastionContext| {
-    ///             # let child_ref = children_ref.elems()[0].clone();
-    ///             # async move {
-    /// // Later, the message is "asked" to the child...
-    /// let answer: Answer = child_ref.ask_anonymously(ASK_MSG).expect("Couldn't send the message.");
-    ///
-    /// // ...and the child's answer is received...
-    /// msg! { answer.await.expect("Couldn't receive the answer."),
-    ///     msg: &'static str => {
-    ///         assert_eq!(msg, ANSWER_MSG);
-    ///         // Handle the answer...
-    ///     };
-    ///     // This won't happen because this example
-    ///     // only answers a `&'static str`...
-    ///     _: _ => ();
-    /// }
-    ///                 #
-    ///                 # Ok(())
-    ///             # }
-    ///         # })
-    ///     # }).unwrap();
-    ///     #
-    ///     # Bastion::start();
-    ///     # Bastion::stop();
-    ///     # Bastion::block_until_stopped();
-    /// # }
-    /// ```
-    ///
-    /// [`Answer`]: message/struct.Answer.html
-    pub fn ask_anonymously<M: Message>(&self, msg: M) -> Result<Answer, M> {
-        debug!("ChildRef({}): Asking message: {:?}", self.id(), msg);
-        let (msg, answer) = BastionMessage::ask(msg);
-        let env = Envelope::from_dead_letters(msg);
-        // FIXME: panics?
-        self.send(env).map_err(|env| env.into_msg().unwrap())?;
-
-        Ok(answer)
-    }
-
-    /// Sends a message to the child this `ChildRef` is referencing
-    /// to tell it to stop its execution.
-    ///
-    /// This method returns `()` if it succeeded, or `Err(())`
-    /// otherwise.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use bastion::prelude::*;
-    /// #
-    /// # fn main() {
-    ///     # Bastion::init();
-    ///     #
-    ///     # let children_ref = Bastion::children(|children| children).unwrap();
-    ///     # let child_ref = &children_ref.elems()[0];
-    /// child_ref.stop().expect("Couldn't send the message.");
-    ///     #
-    ///     # Bastion::start();
-    ///     # Bastion::stop();
-    ///     # Bastion::block_until_stopped();
-    /// # }
-    /// ```
-    pub fn stop(&self) -> Result<(), ()> {
-        debug!("ChildRef({}): Stopping.", self.id);
-        let msg = BastionMessage::stop();
-        let env = Envelope::from_dead_letters(msg);
-        self.send(env).map_err(|_| ())
-    }
-
-    /// Sends a message to the child this `ChildRef` is referencing
-    /// to tell it to suicide.
-    ///
-    /// This method returns `()` if it succeeded, or `Err(())`
-    /// otherwise.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use bastion::prelude::*;
-    /// #
-    /// # fn main() {
-    ///     # Bastion::init();
-    ///     #
-    ///     # let children_ref = Bastion::children(|children| children).unwrap();
-    ///     # let child_ref = &children_ref.elems()[0];
-    /// child_ref.kill().expect("Couldn't send the message.");
-    ///     #
-    ///     # Bastion::start();
-    ///     # Bastion::stop();
-    ///     # Bastion::block_until_stopped();
-    /// # }
-    /// ```
-    pub fn kill(&self) -> Result<(), ()> {
-        debug!("ChildRef({}): Killing.", self.id());
-        let msg = BastionMessage::kill();
-        let env = Envelope::from_dead_letters(msg);
-        self.send(env).map_err(|_| ())
-    }
-
-    /// Returns [`RefAddr`] for the child
-    pub fn addr(&self) -> RefAddr {
-        RefAddr::new(self.path.clone(), self.sender.clone())
-    }
-
-    pub(crate) fn send(&self, env: Envelope) -> Result<(), Envelope> {
-        trace!("ChildRef({}): Sending message: {:?}", self.id(), env);
-        self.sender
-            .unbounded_send(env)
-            .map_err(|err| err.into_inner())
-    }
-
-    pub(crate) fn sender(&self) -> &Sender {
-        &self.sender
-    }
-
-    pub(crate) fn path(&self) -> &Arc<BastionPath> {
-        &self.path
-    }
-}
-
-impl Future for Exec {
-    type Output = Result<(), ()>;
-
-    fn poll(self: Pin<&mut Self>, ctx: &mut Context) -> Poll<Self::Output> {
-        Pin::new(&mut self.get_mut().0).poll(ctx)
-    }
-}
-
-impl Default for Init {
-    fn default() -> Self {
-        Init::new(|_| async { Ok(()) })
-    }
-}
-
-impl PartialEq for ChildrenRef {
-    fn eq(&self, other: &Self) -> bool {
-        self.id == other.id
-    }
-}
-
-impl PartialEq for ChildRef {
-    fn eq(&self, other: &Self) -> bool {
-        self.id == other.id
-    }
-}
-
-impl Eq for ChildrenRef {}
-impl Eq for ChildRef {}
-
-impl Debug for Init {
-    fn fmt(&self, fmt: &mut Formatter) -> fmt::Result {
-        fmt.debug_struct("Init").finish()
-    }
-}
-
-impl Debug for Exec {
-    fn fmt(&self, fmt: &mut Formatter) -> fmt::Result {
-        fmt.debug_struct("Exec").finish()
     }
 }
